backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
celery==5.3.6
redis==5.0.4
yt-dlp>=2025.3.31
openai>=1.60.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
httpx>=0.28.0
aiofiles==23.2.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
requests>=2.31.0
ultralytics>=8.3.0
mediapipe==0.10.14
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
google-api-python-client==2.131.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
cryptography==42.0.8
google-genai>=1.0.0
